+++ src/TrigStation.cpp
@@ -2032,7 +2032,7 @@ struct TrigStation : SickoTrigStation {
 					divClockSample[t] = 1.0;
 					divMaxSample[t][0] = 0.0;
 					divMaxSample[t][1] = 0.0;
-					divCount[t] = 1;
+					divCount[t] = 999999;
 					//outputs[DIVMULT_OUTPUT+t].setVoltage(0.f);
 					//outputs[OUT_OUTPUT+t].setVoltage(0.f);
 
@@ -2061,7 +2061,7 @@ struct TrigStation : SickoTrigStation {
 					divClockSample[t] = 1.0;
 					divMaxSample[t][0] = 0.0;
 					divMaxSample[t][1] = 0.0;
-					divCount[t] = 1;
+					divCount[t] = 999999;
 					//outputs[DIVMULT_OUTPUT+t].setVoltage(0.f);
 					//outputs[OUT_OUTPUT+t].setVoltage(0.f);
 					edge[t] = false;
@@ -2117,7 +2117,7 @@ struct TrigStation : SickoTrigStation {
 						divClockSample[t] = 1.0;
 						divMaxSample[t][0] = 0.0;
 						divMaxSample[t][1] = 0.0;
-						divCount[t] = 1;
+						divCount[t] = 999999;
 						//outputs[DIVMULT_OUTPUT+t].setVoltage(0.f);
 						//outputs[OUT_OUTPUT+t].setVoltage(0.f);
 						edge[t] = false;
@@ -2365,7 +2365,7 @@ struct TrigStation : SickoTrigStation {
 					
 					extBeat[MC] = true;
 
-					extClockSample = 1.0;
+					//extClockSample = 1.0;
 					
 
 					// ----------------------------------------EXTERNAL MAIN CLOCK CV -> RE-INIT TRACKS