+++ changelog.md
@@ -1,5 +1,6 @@
 ### 2.7.4 (2025-06-20)
 - added stepStation and trigStation modules
+- added cvMeter module
 - stepSeq8x / trigSeq8x: adjusted RECL and STOR buttons placement
 - all sampler modules: improved memory allocation management
 - slewer/slewer mini: inverted shapes on shape knob